refactored naming of folders and files
