--- src/components/MAIN_CONTAINER.js
@@ -1,7 +1,0 @@
-import TITLE from "./TITLE_DIV"
-import TOP_DIV from "./TOP_COMPONENTS/TOP_DIV"
-import BOTTOM_DIV from "./BOTTOM_COMPONENTS/BOTTOM_DIV"
-export default function MAIN_CONTAINER(){
-    return(<div><TITLE/><div><TOP_DIV/></div>
-    <div><BOTTOM_DIV/></div></div>)
-}
