+++ src/components/MainContainer.js
@@ -1,0 +1,7 @@
+import TITLE from "./TitleDiv"
+import TOP_DIV from "./top-components/TopDiv"
+import BOTTOM_DIV from "./bottom-components/BottomDiv"
+export default function MAIN_CONTAINER(){
+    return(<div><TITLE/><div><TOP_DIV/></div>
+    <div><BOTTOM_DIV/></div></div>)
+}
