+++ src/components/TitleDiv.js
@@ -1,4 +1,4 @@
-import '../components_styling/STYLING_DIV.css'
+import '../components-styling/StylingDiv.css'
 export default function TITLE(){
     return (<center><div className="Title"><h1>A11yMaps</h1></div></center>);
 }
