+++ src/components/bottom-components/BottomDiv.js
@@ -1,5 +1,5 @@
-import LEFT_BOTTOM_CONTAINER from "./LEFT_BOTTOM_CONTAINER"
-import RIGHT_BOTTOM_CONTAINER from "./RIGHT_BOTTOM_CONTAINER"
+import LEFT_BOTTOM_CONTAINER from "./LeftBottomContainer"
+import RIGHT_BOTTOM_CONTAINER from "./RightBottomContainer"
 export default function TOP_DIV(){
     return (<div id="BOTTOM_DIV" style={{display: "flex"}}><LEFT_BOTTOM_CONTAINER/><RIGHT_BOTTOM_CONTAINER/></div>);
 }
