+++ src/components/bottom-components/Details.js
