+++ src/components/bottom-components/RightBottomContainer.js
@@ -1,4 +1,4 @@
-import EVALUATION from "./EVALUATION";
+import EVALUATION from "./Evaluation";
 export default function RIGHT_BOTTOM_CONTAINER(){
     return (<span id="RIGHT_BOTTOM" style={{"display":"inline-block"}}><EVALUATION/></span>);
 }
