+++ src/components/top-components/TopDiv.js
@@ -1,5 +1,5 @@
-import LEFT_TOP_CONTAINER from "./LEFT_TOP_CONTAINER"
-import RIGHT_TOP_CONTAINER from "./RIGHT_TOP_CONTAINER"
+import LEFT_TOP_CONTAINER from "./LeftTopContainer"
+import RIGHT_TOP_CONTAINER from "./RightTopContainer"
 export default function TOP_DIV(){
     return (<div id="TOP_DIV" style={{display: "flex"}}>
         <LEFT_TOP_CONTAINER/>
